Response status code for the POST, DELETE and PUT
All the response will return 200 as the codes are executed on local machine and since no element is created post the POST call so the ideal outcome will be just checking whether the call was made to that particular method or not so the status code for all the methods: GET, POST, PUT, DELETE will be 200.
This holds true only for the endpoint test file created in realtime they can relate with appropriate status code viz. GET, PUT,DELETE: 200; POST: 201 for successful result.